model/requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
langchain
langchain-community
langchain-google-genai